# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/authenticate": {
            "post": {
                "description": "Authenticates the user and returns a JWT token if the credentials are valid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user",
                "parameters": [
                    {
                        "description": "User credentials",
                        "name": "AuthenticateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.AuthenticateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.AuthenticateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized or User not found",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications": {
            "post": {
                "description": "Creates a notification and maps it to channels and templates",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Create a notification",
                "parameters": [
                    {
                        "description": "Notification details",
                        "name": "NotificationRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.NotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.NotificationResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/notifications/send": {
            "post": {
                "description": "Sends a notification to a user based on the notification ID and username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Send a notification to a user",
                "parameters": [
                    {
                        "description": "Notification details",
                        "name": "SendNotificationRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SendNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/templates": {
            "post": {
                "description": "Creates a template for a specific notification channel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Templates"
                ],
                "summary": "Create a notification template",
                "parameters": [
                    {
                        "description": "Template details",
                        "name": "CreateTemplateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.TemplateResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/user-notification-channels": {
            "post": {
                "description": "Configures notification channels for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Notifications"
                ],
                "summary": "Configure user notification channels",
                "parameters": [
                    {
                        "description": "User notification channel details",
                        "name": "UserNotificationChannelsRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UserNotificationChannelsRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.UserNotificationChannelsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Request",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/common.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ChannelToTemplate": {
            "type": "object",
            "properties": {
                "channel": {
                    "type": "string"
                },
                "templateID": {
                    "type": "string"
                }
            }
        },
        "api.CreateTemplateRequest": {
            "type": "object",
            "properties": {
                "channel": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                }
            }
        },
        "api.NotificationRequest": {
            "type": "object",
            "properties": {
                "channelToTemplateID": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ChannelToTemplate"
                    }
                },
                "priority": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api.NotificationResponse": {
            "type": "object",
            "properties": {
                "channelToTemplateID": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ChannelToTemplate"
                    }
                },
                "id": {
                    "type": "string"
                },
                "priority": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "api.SendNotificationRequest": {
            "type": "object",
            "properties": {
                "notificationID": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.TemplateResponse": {
            "type": "object",
            "properties": {
                "channel": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "template": {
                    "type": "string"
                }
            }
        },
        "api.UserNotificationChannelsRequest": {
            "type": "object",
            "properties": {
                "channels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "notificationID": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.UserNotificationChannelsResponse": {
            "type": "object",
            "properties": {
                "channels": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "notificationID": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.AuthenticateRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "auth.AuthenticateResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}